Fix toggle
