+++ scripts/core/menu/menu.ts
@@ -3,7 +3,7 @@ import {GlobalMenuHorizontal} from './GlobalMenuHorizontal';
 import {appConfig} from 'appConfig';
 import {addInternalEventListener} from 'core/internal-events';
 import {IFullWidthPageCapabilityConfiguration} from 'superdesk-api';
-import {NewAuthoringToggle} from './new-authoring-toggle';
+import {AuthoringToggle} from './authoring-toggle';
 
 SuperdeskFlagsService.$inject = [];
 function SuperdeskFlagsService() {
@@ -76,11 +76,11 @@ angular.module('superdesk.core.menu', [
     .component(
         'sdNewAuthoringToggle',
         reactToAngular1(
-            NewAuthoringToggle,
+            AuthoringToggle,
             [],
         ),
     )
-    
+
 
     // set flags for other directives
     .directive('sdSuperdeskView', ['asset', function(asset) {
